% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTokensByPos.R
\name{countTokensByPos}
\alias{countTokensByPos}
\title{Count Tokens by POS}
\usage{
countTokensByPos(texts, ...)
}
\arguments{
\item{texts}{character. A vector of texts.}

\item{...}{parameters to be pased to \code{pos_tokens}}
}
\value{
list.
}
\description{
Count Tokens by POS
}
