% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropRedundantTags.R
\name{dropRedundantTags}
\alias{dropRedundantTags}
\title{Tag Dropper}
\usage{
dropRedundantTags(pos_lists, sparse = 0.99)
}
\arguments{
\item{texts}{a list of character vectors}
}
\value{
a list of character vectors
}
\description{
Drop redundant parses
}
