% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTM.R
\name{DTM}
\alias{DTM}
\title{Document Term Matricizer}
\usage{
DTM(
  texts,
  sparse = 0.99,
  wstem = "all",
  ngrams = 1,
  language = "english",
  vocabmatch = NULL,
  stop.words = TRUE,
  punct = FALSE,
  POS = FALSE,
  dependency = FALSE,
  tag.sub = 0,
  overlap = 0.8,
  group.conc = NULL,
  group.conc.cutoff = 0.8,
  TPformat = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{texts}{a character vector of texts.}

\item{TPformat}{logical - return in stm::textProcessor() format?}
}
\value{
Feature counts, as a matrix (or in stm format)
}
\description{
Turns text into data.
}
