% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTokensByPos.R
\name{lookUpTokenPos}
\alias{lookUpTokenPos}
\title{Look up texts by token}
\usage{
lookUpTokenPos(df_token_pos_text, s_token_pos)
}
\value{
character vector of texts that include \code{s_token_pos}
}
\description{
Look up texts by token
}
