% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramTokens.R
\name{ngramTokens}
\alias{ngramTokens}
\title{Ngram Tagger}
\usage{
ngramTokens(
  texts,
  wstem = "all",
  ngrams = 1,
  language = "english",
  punct = TRUE,
  stop.words = TRUE,
  overlap = 1,
  sparse = 0.99,
  verbose = FALSE
)
}
\arguments{
\item{texts}{a character vector of texts.}
}
\value{
a matrix of feature counts
}
\description{
Tally bag-of-words ngram features
}
