% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posTokens.R
\name{posTokens}
\alias{posTokens}
\title{POS Tagger}
\usage{
posTokens(
  texts,
  ngrams = 1,
  language = "english",
  punct = FALSE,
  stop.words = TRUE,
  overlap = 1,
  sparse = 0.99,
  dependency = FALSE,
  tag.sub = 0,
  verbose = FALSE
)
}
\arguments{
\item{texts}{a character vector of texts.}

\item{ngrams}{numeric vector of ngram sizes (max = 1:3)}

\item{language}{character what language are you parsing?}

\item{punct}{logical should exclamation points and question marks be included as features?}

\item{stop.words}{logical should stop words be included? default is TRUE}

\item{overlap}{numeric How dissimilar (in cossine distance) must an ngram be from all (n-1)grams to be added to feature set?}

\item{sparse}{maximum feature sparsity for inclusion (1 = include all features)}

\item{dependency}{logical should features have dependency relations appended? default is FALSE}

\item{tag.sub}{numeric what fraction of features should be replaced by POS tags? default is 0 (no features), fractions not supported yet.}

\item{verbose}{logical - report interim steps during processing}

\item{POS}{logical should features have part of speech tags appended? default is FALSE}
}
\value{
a matrix of feature counts
}
\description{
Tally parse-dependent features
}
