% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagDependency.R
\name{tagDependency}
\alias{tagDependency}
\title{Dependency Tagger}
\usage{
tagDependency(dt_parsedtxt)
}
\arguments{
\item{txt}{a character vector of texts.}
}
\value{
list of dependency-tagged items.
}
\description{
Return dependency tags from natural language.
}
