% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unStemmer.R
\name{unStemmer}
\alias{unStemmer}
\title{The Un-Stemmer}
\usage{
unStemmer(
  vocab,
  texts,
  wstem = "all",
  ngrams = 1,
  language = "english",
  stop.words = TRUE
)
}
\arguments{
\item{vocab}{a character vector of stemmed words.}

\item{texts}{a character vector of texts.}
}
\value{
a matrix of feature counts
}
\description{
Returns root words that match the stems
}
