% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posTokens.R
\name{posTokens}
\alias{posTokens}
\title{POS Tagger}
\usage{
posTokens(
  texts,
  ngrams = 1,
  language = "english",
  punct = FALSE,
  stop.words = TRUE,
  overlap = 1,
  sparse = 0.99,
  dependency = FALSE,
  tag.sub = 0,
  verbose = FALSE
)
}
\arguments{
\item{texts}{a character vector of texts.}
}
\value{
a matrix of feature counts
}
\description{
Tally parse-dependent features
}
